+++ src/middlewares/validation.ts
@@ -1,26 +1,10 @@
 import { RequestHandler } from "express";
-import { IsEmail, MinLength, NotContains, validate } from "class-validator";
+import { validateLogin } from "../common/validators/loginValidator";
+import { validateRegister } from "../common/validators/registerValidator";
 
 const registerVal: RequestHandler = (req, res, next) => {
-  class registerForm {
-    @MinLength(2)
-    name: string;
-
-    @IsEmail()
-    email: string;
-
-    @MinLength(8)
-    @NotContains(" ")
-    password: string;
-  }
-
-  const { name, email, password }: registerForm = req.body;
-  const input = new registerForm();
-  input.name = name;
-  input.email = email;
-  input.password = password.trim();
-
-  validate(input, { validationError: { target: false } }).then((errors) => {
+  const { name, email, password } = req.body;
+  validateRegister(name, email, password).then((errors) => {
     if (errors.length > 0) {
       return res.status(400).json({ errors });
     } else next();
@@ -28,21 +12,8 @@ const registerVal: RequestHandler = (req, res, next) => {
 };
 
 const loginVal: RequestHandler = (req, res, next) => {
-  class loginForm {
-    @IsEmail()
-    email: string;
-
-    @MinLength(8)
-    @NotContains(" ")
-    password: string;
-  }
-
-  const { email, password }: loginForm = req.body;
-  const input = new loginForm();
-  input.email = email;
-  input.password = password.trim();
-
-  validate(input, { validationError: { target: false } }).then((errors) => {
+  const { email, password } = req.body;
+  validateLogin(email, password).then((errors) => {
     if (errors.length > 0) {
       return res.status(400).json({ errors });
     } else next();
